+++ src/atoms/BoltArrowUpIcon/BoltArrowUpIcon.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+export default function BoltArrowUpIcon() {
+	return (
+		<svg xmlns="http://www.w3.org/2000/svg" width="17" height="22" viewBox="0 0 17 22"><path fill="currentColor" d="M4,22H3l1-7H.5c-.88,0-.33-.75-.31-.78Q2.13,10.8,6,4H7L6,11H9.52c.4,0,.62.19.4.66Z"/><path fill="currentColor"d="M11.62,0V1.52H14.4L10,5.92,11.08,7l4.4-4.4V5.38H17V0Z"/></svg>
+	);
+}