+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2021-04-14
+### Added
+- Electricity delivered 8 hours chart widget for dashboard.
+
+### Changed
+- Bolt icons for delivered and received.
+- Widget titles for delivered and received electricity.
+
 ## [1.2.0] - 2021-04-13
 ### Added
 - About settings page with app en server version numbers.